pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "olc-downloader"
version = "0.1.0"
description = "CLI tool to download IGC files from onlinecontest.org"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Thomas van der Velden"}
]
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0",
    "scrapy>=2.11.0",
    "scrapy-playwright>=0.0.34",
    "folium>=0.14.0",
]

[project.scripts]
olc-download = "olc_downloader.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
r2 = [
    "boto3>=1.34.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"
